Refactor scanner to use ROWS and COLS as constants.
- DiodeConfiguration now has ROWS and COLS const generics.

- DiodeConfiguration defines two constants, WRITE_LINES and READ_LINES,
  defining how many write/read lines it requires. These are set from the
ROWS and COLS parameters, to tell the scanner which is which.

- ScanMatrix uses DiodeConfiguration internally to convert between
  physical key positions and scan positions. We can also use the
associated constants WRITE_LINES and READ_LINES for the bounds on
`W: WriteLines` and `R: ReadLines`.

Now that the scanner is refactored to publically use the ROWS and COLS constants
and derive WRITE_LINES and READ_LINES internally, it is much easier to
use type checking to verify that it is compatible with other parts of
the system, including making sure that the keymap matches in size.